Selenoid test run. Main tests file temporary ignored, pytest.ini added to VCS.
