+++ pytest.ini
@@ -1,0 +1,4 @@
+[pytest]
+addopts =
+    --clean-alluredir
+    --alluredir=allure-results
